deep_research/pyproject.toml
[tool.setuptools.packages.find]
include = ["notebook_agent*", "research_agent*", "law_agent*"]

[project]
name = "deep-research-example"
version = "0.1.0"
requires-python = ">=3.11, <3.14" # It is recommended to keep 3.12+ for better performance
dependencies = [
    "langchain-openai>=1.1.7",
    "langchain-anthropic>=1.3.1",
    "langchain-tavily>=0.2.16",
    "langgraph-api>=0.7.25",
    "pydantic>=2.12.5",
    "rich>=14.0.0",
    "jupyter>=1.1.0",
    "ipykernel>=6.29.0",
    "tavily-python>=0.5.2",
    "httpx>=0.28.1",
    "markdownify>=1.2.0",
    "deepagents>=0.3.9",
    "python-dotenv>=1.0.1",
    "langgraph-cli[inmem]>=0.4.12",
    "langchain-google-genai>=3.2.0",
    "langchain-mcp-adapters>=0.2.1",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.14.0",
    "ruff>=0.8.4",
]
